refactor: extract hub module-level helpers into hub_helpers.py
